+++ cloudevents/sdk/converters/structured.py
@@ -16,23 +16,24 @@ import typing
 
 from cloudevents.sdk import types
 from cloudevents.sdk.converters import base
+from cloudevents.sdk.converters.util import has_binary_headers
 from cloudevents.sdk.event import base as event_base
 
 
+# TODO: Singleton?
 class JSONHTTPCloudEventConverter(base.Converter):
 
     TYPE = "structured"
     MIME_TYPE = "application/cloudevents+json"
 
     def can_read(
-        self,
-        content_type: str,
-        headers: typing.Dict[str, str] = {"ce-specversion": None},
+        self, content_type: str, headers: typing.Dict[str, str] = {}
     ) -> bool:
         return (
             isinstance(content_type, str)
             and content_type.startswith(self.MIME_TYPE)
-        ) or ("ce-specversion" not in headers)
+            or not has_binary_headers(headers)
+        )
 
     def event_supported(self, event: object) -> bool:
         # structured format supported by both spec 0.1 and 0.2