+++ cloudevents/tests/test_data_encaps_refs.py
@@ -12,15 +12,12 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 
-import copy
-import io
 import json
 from uuid import uuid4
 
 import pytest
 
 from cloudevents.sdk import converters, marshaller
-from cloudevents.sdk.converters import structured
 from cloudevents.sdk.event import v1, v03
 from cloudevents.tests import data
 
@@ -71,7 +68,6 @@ def test_general_binary_properties(event_class):
 
 @pytest.mark.parametrize("event_class", [v03.Event, v1.Event])
 def test_general_structured_properties(event_class):
-    copy_of_ce = copy.deepcopy(data.json_ce[event_class])
     m = marshaller.NewDefaultHTTPMarshaller()
     http_headers = {"content-type": "application/cloudevents+json"}
     event = m.FromRequest(
@@ -92,7 +88,6 @@ def test_general_structured_properties(event_class):
         if key == "content-type":
             assert new_headers[key] == http_headers[key]
             continue
-        assert key in copy_of_ce
 
     # Test setters
     new_type = str(uuid4())