+++ CHANGELOG.md
@@ -4,10 +4,36 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0]
+### Added
+- Added GenericException, DataMarshallingError and DataUnmarshallingError ([#120])
+
+## [1.1.0]
+### Changed
+- Changed from_http to now expect headers argument before data ([#110])
+- Renamed exception names ([#111])
+
+### Fixed
+- Fixed from_http bugs with data of type None, or not dict-like ([#119])
+
+### Deprecated
+- Renamed to_binary_http and to_structured_http. ([#108])
+
+## [1.0.1]
+### Added
+- CloudEvent exceptions and event type checking in http module ([#96])
+- CloudEvent equality override ([#98])
+
 ## [1.0.0]
 ### Added
+- Update types and handle data_base64 structured ([#34])
 - Added a user friendly CloudEvent class with data validation ([#36])
 - CloudEvent structured cloudevent support ([#47])
+- Separated http methods into cloudevents.http module ([#60])
+- Implemented to_json and from_json in http module ([#72])
+
+### Fixed
+- Fixed top level extensions bug ([#71])
 
 ### Removed
 - Removed support for Cloudevents V0.2 and V0.1 ([#43])
@@ -74,6 +100,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#23]: https://github.com/cloudevents/sdk-python/pull/23
 [#25]: https://github.com/cloudevents/sdk-python/pull/25
 [#27]: https://github.com/cloudevents/sdk-python/pull/27
+[#34]: https://github.com/cloudevents/sdk-python/pull/34
 [#36]: https://github.com/cloudevents/sdk-python/pull/36
 [#43]: https://github.com/cloudevents/sdk-python/pull/43
 [#47]: https://github.com/cloudevents/sdk-python/pull/47
+[#60]: https://github.com/cloudevents/sdk-python/pull/60
+[#71]: https://github.com/cloudevents/sdk-python/pull/71
+[#72]: https://github.com/cloudevents/sdk-python/pull/72
+[#96]: https://github.com/cloudevents/sdk-python/pull/96
+[#98]: https://github.com/cloudevents/sdk-python/pull/98
+[#108]: https://github.com/cloudevents/sdk-python/pull/108
+[#110]: https://github.com/cloudevents/sdk-python/pull/110
+[#111]: https://github.com/cloudevents/sdk-python/pull/111
+[#119]: https://github.com/cloudevents/sdk-python/pull/119
+[#120]: https://github.com/cloudevents/sdk-python/pull/120